+++ test/acceptance/turbo-cartocss/error-cases.js
@@ -56,7 +56,7 @@ describe('turbo-carto error cases', function() {
 
             assert.ok(layergroup.hasOwnProperty('errors'));
             assert.equal(layergroup.errors.length, 1);
-            assert.ok(layergroup.errors[0].match(/^turbo-carto/));
+            assert.ok(layergroup.errors[0].match(/^Failed\sto\sprocess/));
             assert.ok(layergroup.errors[0].match(/invalid\sramp\slength/i));
 
             done();
@@ -70,7 +70,7 @@ describe('turbo-carto error cases', function() {
 
             assert.ok(layergroup.hasOwnProperty('errors'));
             assert.equal(layergroup.errors.length, 1);
-            assert.ok(layergroup.errors[0].match(/^turbo-carto/));
+            assert.ok(layergroup.errors[0].match(/^Failed\sto\sprocess/));
             assert.ok(layergroup.errors[0].match(/unable\sto\scompute\sramp/i));
             assert.ok(layergroup.errors[0].match(/wadus_column/));
 
@@ -85,7 +85,7 @@ describe('turbo-carto error cases', function() {
 
             assert.ok(layergroup.hasOwnProperty('errors'));
             assert.equal(layergroup.errors.length, 1);
-            assert.ok(layergroup.errors[0].match(/^turbo-carto/));
+            assert.ok(layergroup.errors[0].match(/^Failed\sto\sprocess/));
             assert.ok(layergroup.errors[0].match(/unable\sto\scompute\sramp/i));
             assert.ok(layergroup.errors[0].match(/invalid\smethod\s\"wadusmethod\"/i));
 
@@ -100,7 +100,7 @@ describe('turbo-carto error cases', function() {
 
             assert.ok(layergroup.hasOwnProperty('errors'));
             assert.equal(layergroup.errors.length, 1);
-            assert.ok(!layergroup.errors[0].match(/^turbo-carto/));
+            assert.ok(!layergroup.errors[0].match(/^Failed\sto\sprocess/));
             assert.ok(layergroup.errors[0].match(/invalid\scode/i));
 
             done();
@@ -116,7 +116,7 @@ describe('turbo-carto error cases', function() {
             assert.equal(layergroup.errors_with_context.length, 1);
             assert.equal(layergroup.errors_with_context[0].type, 'layer');
             assert.equal(layergroup.errors_with_context[0].subtype, 'turbo-carto');
-            assert.ok(layergroup.errors_with_context[0].message.match(/^turbo-carto/));
+            assert.ok(layergroup.errors_with_context[0].message.match(/^Failed\sto\sprocess/));
             assert.ok(layergroup.errors_with_context[0].message.match(/unable\sto\scompute\sramp/i));
             assert.ok(layergroup.errors_with_context[0].message.match(/wadus_column/));
 
@@ -172,14 +172,14 @@ describe('turbo-carto error cases', function() {
 
             assert.equal(layergroup.errors_with_context[0].type, 'layer');
             assert.equal(layergroup.errors_with_context[0].subtype, 'turbo-carto');
-            assert.ok(layergroup.errors_with_context[0].message.match(/^turbo-carto/));
+            assert.ok(layergroup.errors_with_context[0].message.match(/^Failed\sto\sprocess/));
             assert.ok(layergroup.errors_with_context[0].message.match(/unable\sto\scompute\sramp/i));
             assert.ok(layergroup.errors_with_context[0].message.match(/wadus_column/));
             assert.equal(layergroup.errors_with_context[0].layer.id, 'layer0');
 
             assert.equal(layergroup.errors_with_context[1].type, 'layer');
             assert.equal(layergroup.errors_with_context[1].subtype, 'turbo-carto');
-            assert.ok(layergroup.errors_with_context[1].message.match(/^turbo-carto/));
+            assert.ok(layergroup.errors_with_context[1].message.match(/^Failed\sto\sprocess/));
             assert.ok(layergroup.errors_with_context[1].message.match(/unable\sto\scompute\sramp/i));
             assert.ok(layergroup.errors_with_context[1].message.match(/invalid_column/));
             assert.equal(layergroup.errors_with_context[1].layer.id, 'layer1');